# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Enterprise ROI Automation Platform API
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export type ProjectStatus = (typeof ProjectStatus)[keyof typeof ProjectStatus];

export const ProjectStatus = {
  draft: "draft",
  review: "review",
  approved: "approved",
  archived: "archived",
} as const;

export interface ProjectSalesRegion {
  id: number;
  projectId: number;
  region: string;
  fxMultiplier: number;
}

export interface Project {
  id: number;
  name: string;
  /** @nullable */
  description?: string | null;
  region: string;
  /** @nullable */
  productCategory?: string | null;
  /** @nullable */
  investmentSize?: number | null;
  status: ProjectStatus;
  /** @nullable */
  createdBy?: string | null;
  createdAt: string;
  updatedAt: string;
  salesRegions?: ProjectSalesRegion[];
}

export interface ProjectSalesRegionInput {
  region: string;
  fxMultiplier?: number;
}

export interface ProjectSalesRegionUpdate {
  fxMultiplier?: number;
}

export interface ProjectInput {
  /** @minLength 1 */
  name: string;
  description?: string;
  region: string;
  productCategory?: string;
  investmentSize?: number;
  createdBy?: string;
}

export type ProjectUpdateStatus =
  (typeof ProjectUpdateStatus)[keyof typeof ProjectUpdateStatus];

export const ProjectUpdateStatus = {
  draft: "draft",
  review: "review",
  approved: "approved",
  archived: "archived",
} as const;

export interface ProjectUpdate {
  /** @minLength 1 */
  name?: string;
  description?: string;
  region?: string;
  productCategory?: string;
  investmentSize?: number;
  status?: ProjectUpdateStatus;
}

export interface Calculation {
  id: number;
  scenarioId: number;
  npv: number;
  /** @nullable */
  irr: number | null;
  /** @nullable */
  paybackPeriod: number | null;
  /** @nullable */
  roiPercent: number | null;
  /** @nullable */
  totalRevenue?: number | null;
  /** @nullable */
  totalCost?: number | null;
  /** @nullable */
  totalCogs?: number | null;
  /** @nullable */
  totalCapex?: number | null;
  /** @nullable */
  cashFlows?: string | null;
  calculatedAt: string;
}

export interface Scenario {
  id: number;
  projectId: number;
  name: string;
  isBaseline: boolean;
  lifecycleYears: number;
  /** @nullable */
  assumptionsSummary?: string | null;
  createdAt: string;
  calculation?: Calculation;
}

export interface ScenarioInput {
  /** @minLength 1 */
  name: string;
  /**
   * @minimum 1
   * @maximum 10
   */
  lifecycleYears?: number;
  assumptionsSummary?: string;
  isBaseline?: boolean;
}

export interface ScenarioUpdate {
  name?: string;
  /**
   * @minimum 1
   * @maximum 10
   */
  lifecycleYears?: number;
  assumptionsSummary?: string;
  isBaseline?: boolean;
}

export interface DuplicateScenarioInput {
  /** @minLength 1 */
  name: string;
}

export interface FinancialInput {
  id: number;
  scenarioId: number;
  year: number;
  salesVolume: number;
  netPrice: number;
  /** @nullable */
  unitCost?: number | null;
  /** @nullable */
  costQty?: number | null;
  cogs: number;
  opex: number;
  capex: number;
}

export type FinancialInputsBulkInputsItem = {
  year: number;
  salesVolume: number;
  netPrice: number;
  /** @nullable */
  unitCost?: number | null;
  /** @nullable */
  costQty?: number | null;
  cogs: number;
  opex: number;
  capex: number;
};

export interface FinancialInputsBulk {
  inputs: FinancialInputsBulkInputsItem[];
}

export interface RegionalParameter {
  id: number;
  region: string;
  taxRate: number;
  currency: string;
  exchangeRateToUsd: number;
  effectiveDate: string;
}

export interface RegionalParameterInput {
  region: string;
  taxRate: number;
  currency: string;
  exchangeRateToUsd: number;
  effectiveDate: string;
}

export interface RegionalParameterUpdate {
  taxRate?: number;
  currency?: string;
  exchangeRateToUsd?: number;
  effectiveDate?: string;
}

export interface FormulaDefinition {
  id: number;
  formulaType: string;
  wacc: number;
  discountRate: number;
  /** @nullable */
  parametersJson?: string | null;
  locked: boolean;
  version: number;
  /** @nullable */
  updatedBy?: string | null;
  updatedAt: string;
}

export interface FormulaDefinitionInput {
  formulaType: string;
  wacc: number;
  discountRate: number;
  parametersJson?: string;
  locked?: boolean;
}

export interface FormulaDefinitionUpdate {
  wacc?: number;
  discountRate?: number;
  parametersJson?: string;
  locked?: boolean;
}

export interface AuditLogEntry {
  id: number;
  tableName: string;
  recordId: number;
  fieldName: string;
  /** @nullable */
  oldValue: string | null;
  /** @nullable */
  newValue: string | null;
  changedBy: string;
  changedAt: string;
}

export interface SheetsExportResult {
  inputRows: number;
  outputRows: number;
  sheetUrl: string;
}

export interface SheetsImportResult {
  updated: number;
  calculated: number;
  errors: string[];
}

export type DashboardSummaryByRegionItem = {
  region?: string;
  count?: number;
  /** @nullable */
  avgNpv?: number | null;
};

export type DashboardSummaryByStatusItem = {
  status?: string;
  count?: number;
};

export interface DashboardSummary {
  totalProjects: number;
  totalApproved: number;
  totalInReview: number;
  totalDraft: number;
  /** @nullable */
  avgNpv: number | null;
  /** @nullable */
  avgIrr: number | null;
  /** @nullable */
  totalInvestment: number | null;
  byRegion: DashboardSummaryByRegionItem[];
  byStatus: DashboardSummaryByStatusItem[];
}

export interface DashboardProject {
  id: number;
  name: string;
  region: string;
  /** @nullable */
  productCategory?: string | null;
  status: string;
  /** @nullable */
  investmentSize: number | null;
  /** @nullable */
  npv: number | null;
  /** @nullable */
  irr: number | null;
  /** @nullable */
  paybackPeriod: number | null;
  /** @nullable */
  roiPercent: number | null;
  /** @nullable */
  pi?: number | null;
  /** @nullable */
  totalRevenue?: number | null;
  /** @nullable */
  cm?: number | null;
  /** @nullable */
  cmPct?: number | null;
  salesRegions?: string[];
  createdAt: string;
}

export type ListProjectsParams = {
  region?: string;
  status?: string;
  search?: string;
};

export type GetDashboardProjectsParams = {
  region?: string;
  status?: string;
  sortBy?: string;
  sortOrder?: string;
};