# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: Enterprise ROI Automation Platform API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /regional-parameters:
    get:
      operationId: listRegionalParameters
      tags: [regional]
      summary: List all regional parameters
      responses:
        "200":
          description: List of regional parameters
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/RegionalParameter"
    post:
      operationId: createRegionalParameter
      tags: [regional]
      summary: Create a regional parameter entry
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/RegionalParameterInput"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/RegionalParameter"
  /regional-parameters/{id}:
    patch:
      operationId: updateRegionalParameter
      tags: [regional]
      summary: Update a regional parameter
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/RegionalParameterUpdate"
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/RegionalParameter"
  /formula-definitions:
    get:
      operationId: listFormulaDefinitions
      tags: [formulas]
      summary: List formula definitions
      responses:
        "200":
          description: Formula definitions
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/FormulaDefinition"
    post:
      operationId: createFormulaDefinition
      tags: [formulas]
      summary: Create a formula definition
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/FormulaDefinitionInput"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/FormulaDefinition"
  /formula-definitions/{id}:
    patch:
      operationId: updateFormulaDefinition
      tags: [formulas]
      summary: Update a formula definition (admin only)
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/FormulaDefinitionUpdate"
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/FormulaDefinition"
...
  /dashboard/projects:
    get:
      operationId: getDashboardProjects
      tags: [dashboard]
      summary: Get all projects ranked by NPV for executive dashboard
      parameters:
        - name: region
          in: query
          schema:
            type: string
        - name: status
          in: query
          schema:
            type: string
        - name: sortBy
          in: query
          schema:
            type: string
        - name: sortOrder
          in: query
          schema:
            type: string
      responses:
        "200":
          description: Ranked projects
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/DashboardProject"
...
components:
  schemas:
...
    RegionalParameter:
      type: object
      required: [id, region, taxRate, currency, exchangeRateToUsd, effectiveDate]
      properties:
        id:
          type: integer
        region:
          type: string
        taxRate:
          type: number
        currency:
          type: string
        exchangeRateToUsd:
          type: number
        effectiveDate:
          type: string
          format: date
    RegionalParameterInput:
      type: object
      required: [region, taxRate, currency, exchangeRateToUsd, effectiveDate]
      properties:
        region:
          type: string
        taxRate:
          type: number
        currency:
          type: string
        exchangeRateToUsd:
          type: number
        effectiveDate:
          type: string
    RegionalParameterUpdate:
      type: object
      properties:
        taxRate:
          type: number
        currency:
          type: string
        exchangeRateToUsd:
          type: number
        effectiveDate:
          type: string
    FormulaDefinition:
      type: object
      required: [id, formulaType, wacc, discountRate, locked, version, updatedAt]
      properties:
        id:
          type: integer
        formulaType:
          type: string
        wacc:
          type: number
        discountRate:
          type: number
        parametersJson:
          type: ["string", "null"]
        locked:
          type: boolean
        version:
          type: integer
        updatedBy:
          type: ["string", "null"]
        updatedAt:
          type: string
          format: date-time
    FormulaDefinitionInput:
      type: object
      required: [formulaType, wacc, discountRate]
      properties:
        formulaType:
          type: string
        wacc:
          type: number
        discountRate:
          type: number
        parametersJson:
          type: string
        locked:
          type: boolean
    FormulaDefinitionUpdate:
      type: object
      properties:
        wacc:
          type: number
        discountRate:
          type: number
        parametersJson:
          type: string
        locked:
          type: boolean
...
    DashboardProject:
      type: object
      required: [id, name, region, status, investmentSize, npv, irr, paybackPeriod, roiPercent, createdAt]
      properties:
        id:
          type: integer
        name:
          type: string
        region:
          type: string
        productCategory:
          type: ["string", "null"]
        status:
          type: string
        investmentSize:
          type: ["number", "null"]
        npv:
          type: ["number", "null"]
        irr:
          type: ["number", "null"]
        paybackPeriod:
          type: ["number", "null"]
        roiPercent:
          type: ["number", "null"]
        pi:
          type: ["number", "null"]
        totalRevenue:
          type: ["number", "null"]
        cm:
          type: ["number", "null"]
        cmPct:
          type: ["number", "null"]
        salesRegions:
          type: array
          items:
            type: string
        createdAt:
          type: string
          format: date-time