# lib/api-zod/src/generated/types/dashboardProject.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Enterprise ROI Automation Platform API
 * OpenAPI spec version: 0.1.0
 */

export interface DashboardProject {
  id: number;
  name: string;
  region: string;
  /** @nullable */
  productCategory?: string | null;
  status: string;
  /** @nullable */
  investmentSize: number | null;
  /** @nullable */
  npv: number | null;
  /** @nullable */
  irr: number | null;
  /** @nullable */
  paybackPeriod: number | null;
  /** @nullable */
  roiPercent: number | null;
  /** @nullable */
  pi?: number | null;
  /** @nullable */
  totalRevenue?: number | null;
  /** @nullable */
  cm?: number | null;
  /** @nullable */
  cmPct?: number | null;
  salesRegions?: string[];
  createdAt: Date;
}