# lib/api-zod/src/generated/types/formulaDefinitionUpdate.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Enterprise ROI Automation Platform API
 * OpenAPI spec version: 0.1.0
 */

export interface FormulaDefinitionUpdate {
  wacc?: number;
  discountRate?: number;
  parametersJson?: string;
  locked?: boolean;
}