# lib/api-zod/src/generated/types/regionalParameterInput.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Enterprise ROI Automation Platform API
 * OpenAPI spec version: 0.1.0
 */

export interface RegionalParameterInput {
  region: string;
  taxRate: number;
  currency: string;
  exchangeRateToUsd: number;
  effectiveDate: string;
}